+++ mobile/alix/lib/src/rust/frb_generated.dart
@@ -1739,15 +1739,16 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
   InlineRun dco_decode_inline_run(dynamic raw) {
     // Codec=Dco (DartCObject based), see doc to use other codecs
     final arr = raw as List<dynamic>;
-    if (arr.length != 6)
-      throw Exception('unexpected arr length: expect 6 but see ${arr.length}');
+    if (arr.length != 7)
+      throw Exception('unexpected arr length: expect 7 but see ${arr.length}');
     return InlineRun(
       text: dco_decode_String(arr[0]),
       bold: dco_decode_bool(arr[1]),
       italic: dco_decode_bool(arr[2]),
       strike: dco_decode_bool(arr[3]),
       code: dco_decode_bool(arr[4]),
-      math: dco_decode_opt_box_autoadd_math_view(arr[5]),
+      link: dco_decode_bool(arr[5]),
+      math: dco_decode_opt_box_autoadd_math_view(arr[6]),
     );
   }
 
@@ -2723,6 +2724,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     var var_italic = sse_decode_bool(deserializer);
     var var_strike = sse_decode_bool(deserializer);
     var var_code = sse_decode_bool(deserializer);
+    var var_link = sse_decode_bool(deserializer);
     var var_math = sse_decode_opt_box_autoadd_math_view(deserializer);
     return InlineRun(
       text: var_text,
@@ -2730,6 +2732,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
       italic: var_italic,
       strike: var_strike,
       code: var_code,
+      link: var_link,
       math: var_math,
     );
   }
@@ -3934,6 +3937,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     sse_encode_bool(self.italic, serializer);
     sse_encode_bool(self.strike, serializer);
     sse_encode_bool(self.code, serializer);
+    sse_encode_bool(self.link, serializer);
     sse_encode_opt_box_autoadd_math_view(self.math, serializer);
   }